+++ AnimalsNeedWater/Core/Multiplayer/Models/AddFullAnimalMessage.cs
@@ -1,6 +1,4 @@
-using System.Collections.Generic;
-
-namespace AnimalsNeedWater.Core.Models;
+namespace AnimalsNeedWater.Core.Multiplayer.Models;
 
 public class AddFullAnimalMessage
 {
@@ -10,5 +8,4 @@ public class AddFullAnimalMessage
     {
         AnimalId = animalId;
     }
-
 }