+++ AnimalsNeedWater/Core/Multiplayer/Models/StateSyncMessage.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace AnimalsNeedWater.Core.Models;
+namespace AnimalsNeedWater.Core.Multiplayer.Models;
 
 public class StateSyncMessage
 {